@@ -1,6 +1,6 @@
 {
   "name": "ib-saver",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "license": "MIT",
   "scripts": {
     "dev": "next",
@@ -10,28 +10,28 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@directus/sdk": "^10.3.0",
+    "@directus/sdk": "^10.3.1",
     "@fortawesome/fontawesome-pro": "^6.2.1",
-    "axios": "^1.2.0",
     "clsx": "^1.2.1",
     "fathom-client": "^3.5.0",
-    "next": "^13.0.5",
-    "next-seo": "^5.14.1",
+    "next": "^13.1.1",
+    "next-seo": "^5.15.0",
     "next-themes": "^0.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "swr": "^2.0.0",
     "zustand": "^4.1.5"
   },
   "devDependencies": {
     "@next/bundle-analyzer": "^13.1.1",
     "@svgr/webpack": "^6.5.1",
-    "@types/node": "^18.11.9",
-    "@types/react": "^18.0.25",
+    "@types/node": "^18.11.17",
+    "@types/react": "^18.0.26",
     "autoprefixer": "^10.4.13",
-    "eslint": "8.28.0",
-    "eslint-config-next": "13.0.5",
-    "postcss": "^8.4.19",
+    "eslint": "8.30.0",
+    "eslint-config-next": "13.1.1",
+    "postcss": "^8.4.20",
     "tailwindcss": "^3.2.4",
-    "typescript": "^4.9.3"
+    "typescript": "^4.9.4"
   }
 }
